+++ projects/Mallard/src/screens/article-screen.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, ReactNode } from 'react'
+import React, { useState } from 'react'
 import { useJsonOrEndpoint } from 'src/hooks/use-fetch'
 import { NavigationScreenProp, NavigationEvents } from 'react-navigation'
 import {
@@ -8,15 +8,7 @@ import {
 } from 'src/theme/appearance'
 import { Article } from 'src/components/article'
 import { Article as ArticleType, Collection } from 'src/common'
-import {
-    View,
-    TouchableOpacity,
-    Animated,
-    Dimensions,
-    Easing,
-    Platform,
-    StyleSheet,
-} from 'react-native'
+import { View, TouchableOpacity } from 'react-native'
 import { metrics } from 'src/theme/spacing'
 import { UiBodyCopy } from 'src/components/styled-text'
 import { SlideCard } from 'src/components/layout/slide-card/index'
@@ -25,8 +17,8 @@ import { PathToArticle } from './article-screen'
 import { withResponse } from 'src/hooks/use-response'
 import { FlexErrorMessage } from 'src/components/layout/errors/flex-error-message'
 import { ERR_404_REMOTE, ERR_404_MISSING_PROPS } from 'src/helpers/words'
-import MaskedView from '@react-native-community/masked-view'
 import { Issue } from '../../../backend/common'
+import { Clipper } from 'src/components/layout/clipper/clipper'
 
 export interface PathToArticle {
     collection: Collection['key']
@@ -63,69 +55,6 @@ const useArticleResponse = ({ collection, article, issue }: PathToArticle) => {
     return withResponse<ArticleType>(resp)
 }
 
-const clipperStyles = StyleSheet.create({
-    flex: { flex: 1 },
-    bg: { backgroundColor: 'black' },
-})
-
-const Clipper = ({
-    children,
-    transitionProps,
-}: {
-    children: ReactNode
-    transitionProps?: ArticleTransitionProps
-}) => {
-    /* 
-    This is part of the transition from fronts and it positions 
-    the article where it should on screen
-    */
-    const [windowHeight] = useState(() => Dimensions.get('window').height)
-    const [height] = useState(
-        () =>
-            new Animated.Value(
-                (2 *
-                    ((transitionProps &&
-                        transitionProps.startAtHeightFromFrontsItem) ||
-                        windowHeight)) /
-                    windowHeight,
-            ),
-    )
-    useEffect(() => {
-        Animated.sequence([
-            Animated.delay(50),
-            Animated.timing(height, {
-                toValue: 2,
-                duration: 300,
-                easing: Easing.inOut(Easing.linear),
-                useNativeDriver: true,
-            }),
-        ]).start()
-    }, []) // eslint-disable-line react-hooks/exhaustive-deps
-    /* android struggles animating masks sadface */
-    if (!transitionProps || Platform.OS !== 'ios') return <>{children}</>
-    return (
-        <MaskedView
-            style={clipperStyles.flex}
-            maskElement={
-                <Animated.View
-                    style={[
-                        {
-                            transform: [
-                                { translateY: windowHeight / -2 },
-                                { scaleY: height },
-                            ],
-                        },
-                        clipperStyles.flex,
-                        clipperStyles.bg,
-                    ]}
-                />
-            }
-        >
-            {children}
-        </MaskedView>
-    )
-}
-
 const ArticleScreenWithProps = ({
     path,
     articlePrefill,
@@ -151,7 +80,11 @@ const ArticleScreenWithProps = ({
     const [viewIsTransitioning, setViewIsTransitioning] = useState(true)
 
     return (
-        <Clipper {...{ transitionProps }}>
+        <Clipper
+            from={
+                transitionProps && transitionProps.startAtHeightFromFrontsItem
+            }
+        >
             <SlideCard
                 {...viewIsTransitioning}
                 onDismiss={() => navigation.goBack()}
